added delete to notification channels

+++ app/anomaly_detectors/routes.py
@@ -31,6 +31,8 @@ def details(anomaly_detector_id):
 @login_required
 def delete(anomaly_detector_id):
     anomaly_detector = AnomalyDetector.query.get(anomaly_detector_id)
+    if anomaly_detector.user.id != current_user.id:
+        return "", 404
     if anomaly_detector.user_id == current_user.id:
         db.session.delete(anomaly_detector)
         db.session.commit()
